make NbcRotate abstract as it's implementation is noop
During testing the NbcRotate function on SdrObject it turned out
the the function is essentially a noop, because it uses normal
equality to test the values of a double, which are a product of
sin and cos functions (to determine the 90 degree angles). So
because of this the input rectangle was never modified - noop.

Because of this we can just remove the impl. of the function and
declare it abstract, so that the actual implementations define
a valid function to rotate.

There were some subclasses that didn't override the NbcRotate so
they used the one implementation in SdrObject. These subclasses
now override the function and in the implementation we call
assert(false), which is never called during a test run. It seems
we never rotate those objects.

Change-Id: I1b1a45a8e96ed2d061f9b9f80c5fdaa5a84d4c05